Bugfix: save/load model weights did not work
The weights were saved for the original model but were loaded for the cloned
model (in tf_train). In Keras, the cloned model does not have the same layer
structure as the original one. Instead, the whole original model shows up as a
single layer in the cloned model. So weights saved in one cannot be loaded by
another.

The solution is to save and load the weights for the original model, with the
regular layer structure. It's also better because now load/save happens in the
same scope, outside of tf_train.